+++ unipi_control/features/eastron.py
@@ -11,10 +11,10 @@ from pymodbus.payload import BinaryPayloadDecoder
 
 from unipi_control.config import Config
 from unipi_control.config import FeatureConfig
-from unipi_control.features.utils import FeatureType
+from unipi_control.features.constants import FeatureType
 from unipi_control.helpers.text import slugify
-from unipi_control.helpers.typing import HardwareDefinition
-from unipi_control.modbus.helper import ModbusHelper
+from unipi_control.hardware.map import HardwareDefinition
+from unipi_control.modbus.helpers import ModbusHelper
 
 
 @dataclass