+++ unipi_control/hardware/constants.py
@@ -1,0 +1,7 @@
+"""Hardware constants."""
+from typing import Final
+
+
+class HardwareType:
+    PLC: Final[str] = "PLC"
+    EXTENSION: Final[str] = "Extension"