+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for Python 3.12
 - Added support for Unipi Extensions
 - Added a lot of debug log messages for better Modbus RTU/TCP debugging
+- Added `scan_interval` to Modbus RTU/TCP configuration
 
 ### Changed